did some styling realized i have work to do
firt in the popular page. i need to add images(just copy), or i'm
removing it.

then in the checkout page i need to increase decrease the count
along with a remove button

and then i do some testing, just use one or two tests especially
with the ui. Then i would need to do animations. this is overall
a good amount of work